Refactor map IO orchestration
